权限控制

+++ application/index/controller/Comment.php
@@ -11,9 +11,11 @@ use think\Controller;
 
 class Comment extends Controller{
 	public function index(){
-		$list=CommentModel::all();
+		// $list=CommentModel::all();
+		$list=CommentModel::paginate(15);
+		$count=CommentModel::count();
 		$this->assign('list',$list);
-		$this->assign('count',count($list));
+		$this->assign('count',$count);
 		return $this->fetch();
 	}
 	public function input(){
@@ -26,18 +28,36 @@ class Comment extends Controller{
 	}
 	
 	public function edit($id){
+			if(!Session::get('vip')){
+			return '权限不够！';
+		}
+		if(Session::get('vip')->ugroup_id>2){
+			return '权限不够！';
+		}
 		$comment=CommentModel::get($id);
 		$this->assign('comment',$comment);
 		return $this->fetch();
 	}
 	public function update(){
+			if(!Session::get('vip')){
+			return '权限不够！';
+		}
+		if(Session::get('vip')->ugroup_id>2){
+			return '权限不够！';
+		}
 		$comment=CommentModel::get($_POST['id']);
 		$comment->content=$_POST['content'];
 		$category->save();
 		return '修改评论成功';
 	}
 	
-	public function delete($id){
+	public function delete($id){	
+	if(!Session::get('vip')){
+			return '权限不够！';
+		}
+		if(Session::get('vip')->ugroup_id>1){
+			return '权限不够！';
+		}
 		$comment=CommentModel::get($id);
 		$comment->delete();
 		return '评论已删除';
